app/test: support resources externally linked
To include resources from other source that the C source code we can take
advantage of the objcopy behaviour, i.e. packing of an arbitrary file as an
object file that is linked to the target program.

A linked object file is always accessible as a pair

extern const char beg_<name>;
extern const char end_<name>;
(extern const char siz_<name>;)

The objcopy however accepts architecture parameters in a very tricky way.
Try to translate as many arguments as possible from the RTE_ARCH into the
objcopy specific names.

We pack the resource.c source file as an example for testing.

*** CAUTION: ***
The objcopy and resource creation is a subject of change. Any comments of how
to integrate those is welcome.

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>