ixgbe/base: update swfw semaphore function
For X540 and forward it is possible if a system reset occur at the
right time to leave the SWFW semaphore high. This new function will
attempt to grab and release the semaphore. If the grab times out it
will still release the semaphore placing it in a known good state.
The idea is to call this when you know no one should be holding the
semaphore (i.e. probe time)
This patch alse fixes possible race issue between ports when issuing
host interface command by acquiring/releasing the management host
interface semaphore in ixgbe_host_interface_command.

Signed-off-by: Beilei Xing <beilei.xing@intel.com>