i40e: support user unaware VF reset
Problem:
Now the i40e VF PMD driver does not support reset event from PF.
Customer want the user unaware VF reset feature in VF PMD driver,
with this feature, the application doesn't need to concern about
the VF reset procedure and all the application can work correctly
when VF is doing the reset.

Solution:
The following step need to be done
--capture the VF reset event
--reconfigure the VF device
--restore all the configuration parameters.
With this patch, the first two steps can be meet,
part of the configuration parameters can be restored.
This patch mainly focus on the reconfiguration of the VF device when
receive the VF reset event from PF.

Known issues:
This feature cannot work on the primary secondary mixed scenario.

Signed-off-by: zhe.tao <zhe.tao@intel.com>