i40e: fix flexible payload selection
When setting up flexible payload selection rules, it is allowed
that setting value to 63 to disable the rule (NONUSE_FLX_PIT_DEST_OFF).
However, MK_FLX_PIT macro is always adding an offset value 50
(I40E_FLX_OFFSET_IN_FIELD_VECTOR), it will be set to "63 + 50" and
when setting NONUSE_FLX_PIT_DEST_OFF to disable it. It breaks
the functionality.
This patch fixes this issue.

Fixes: d8b90c4eabe9 ("i40e: take flow director flexible payload
                      configuration")

Reported-by: Michael Habibi <mikehabibi@gmail.com>
Signed-off-by: Jingjing Wu <jingjing.wu@intel.com>