tools:new tool for system info CPU, memory and huge pages
The new tool uses /sys/devices instead of /proc directory, which
does not exist on all systems. If the procfs is not available
then memory and huge page information is skipped.

The tool also can emit a json format in short or long form to
allow for machine readable information.

Here is the usage information:

Usage: sys_info.py [options]

Show the lcores layout with core id and socket(s).

Options:
    --help or -h:
        Display the usage information and quit

    --long or -l:
        Display the information in a machine readable long format.

    --short or -s:
        Display the information in a machine readable short format.

    --pci or -p:
        Display all of the Ethernet devices in the system using 'lspci'.

    --version or -v:
        Display the current version number of this script.

    --debug or -d:
        Output some debug information.

    default:
        Display the information in a human readable format.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>