remove rte.hostapp.mk
During the review of this patchset it was discovered that rte.hostapp.mk was
going unused.  All its users had been previously removed, and so this file too
can go.  It and rte.buildtools.mk are effectively duplicates, but given that
pmdinfogen is a tool used to create the dpdk build, I think the naming of the
buildtools variant is more appropriate.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
CC: Bruce Richardson <bruce.richardson@intel.com>
CC: Thomas Monjalon <thomas.monjalon@6wind.com>
CC: Stephen Hemminger <stephen@networkplumber.org>
CC: Panu Matilainen <pmatilai@redhat.com>