bnxt: add HWRM ring alloc/free functions
Add HWRM API calls to allocate and free TX, RX and Completion rings
in the hardware along with the associated structs and definitions.

As mentioned earlier, a completion ring is used by the Ethernet
controller to provide the status of transmitted & received packets,
report errors, status changes to the host software.

v4:
Address review comments.

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Reviewed-by: David Christensen <david.christensen@broadcom.com>
Signed-off-by: Stephen Hurd <stephen.hurd@broadcom.com>