mk: fix parallel build of test resources
The build was failing sometimes when building with multiple
parallel jobs:
    # rm build/build/app/test/*res*
    # make -j6
    objcopy: 'resource.tmp': No such file

The reason is that each resource was built from the same temporary file.
The failure is seen because of a race condition when removing the
temporary file after each resource creation.
It also means that some resources may be created from the wrong source.

The fix is to have a different temporary file for each resource.
We do not use the original source file because it has a long path
which is used by objcopy to name the symbols after some transformations.
The trick is to create a symbolic link of the source file if it is not
already in the current build directory. Then we just have to replace
the dot by an underscore to predict the symbol names to redefine.

Fixes: 1e9e0a6270 ("app/test: fix resource creation with objcopy on FreeBSD")

Signed-off-by: Thomas Monjalon <thomas.monjalon@6wind.com>