drivers: export infos as string symbols
Modify the PMD_REGISTER_DRIVER macro, adding a name argument to it.  The
addition of a name argument creates a token that can be used for subsequent
macros in the creation of unique symbol names to export additional bits of
information for use by the dpdk-pmdinfogen tool.  For example:

    RTE_REGISTER_DRIVER(ena, ena_driver);
registers the ena_driver struct as PMD_REGISTER_DRIVER did, and creates
a symbol
    const char rte_pmd_name0[] __attribute__((used)) = "ena";
which dpdk-pmdinfogen can search for and extract.  The subsequent macro
    RTE_EXPORT_PCI_TABLE(ena, ena_pci_id_map);
creates a symbol
    const char ena_pci_table_export[] __attribute__((used)) = "ena_pci_id_map";
Which allows dpdk-pmdinfogen to find the pci table of this driver.

Using this pattern, we can export arbitrary bits of information.

dpdk-pmdinfo uses this information to extract hardware support from an object
file and create a json string to make hardware support info discoverable
later.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Acked-by: Panu Matilainen <pmatilai@redhat.com>
Acked-by: Remy Horton <remy.horton@intel.com>
Signed-off-by: Thomas Monjalon <thomas.monjalon@6wind.com>