qat: change the session structure to be variable sized
This patch changes the qat session data structure sent to qat from a
fixed size to a variable size which is dependent on the size of
the chosen algorithm.
This reduces the amount of bytes which are transferred across
PCIe and thus helps to increase qat performance when the
accelerator is bound by PCIe.

Signed-off-by: John Griffin <john.griffin@intel.com>